feat: Post 로직 전체 @AuthenticationPrincipal User 가져오기 로직 다시 추가 - merge하는 과정에 덮어써진 파일 다시 불러오기

+++ src/main/java/opensocial/org/community_hub/domain/post/service/PostService.java
@@ -3,6 +3,7 @@ package opensocial.org.community_hub.domain.post.service;
 import lombok.RequiredArgsConstructor;
 import opensocial.org.community_hub.domain.post.entity.Post;
 import opensocial.org.community_hub.domain.post.repository.PostRepository;
+import opensocial.org.community_hub.domain.user.entity.User;
 import org.springframework.stereotype.Service;
 
 import java.util.Optional;
@@ -14,7 +15,9 @@ public class PostService {
     private final PostRepository postRepository;
 
     // 게시글 생성
-    public Post createPost(Post post) {
+    public Post createPost(Post post, User user) {
+
+        post.setUser(user); // 게시글에 사용자 정보 추가
         return postRepository.save(post);
     }
 
@@ -23,23 +26,30 @@ public class PostService {
         return postRepository.findById(postId);
     }
 
-    // 게시글 업데이트
-    public Post updatePost(Long postId, Post postDetails) {
+    // 게시글 업데이트 (본인 게시글만 업데이트)
+    public Post updatePost(Long postId, Post postDetails, User user) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RuntimeException("Post not found with id " + postId));
+
+        if (!post.getUser().getUserId().equals(user.getUserId())) {
+            throw new RuntimeException("You are not authorized to update this post");
+        }
 
         post.setTitle(postDetails.getTitle());
         post.setContent(postDetails.getContent());
         post.setViewCount(postDetails.getViewCount());
-        post.setCommentCount(postDetails.getCommentCount());
-
         return postRepository.save(post);
     }
 
-    // 게시글 삭제
-    public void deletePost(Long postId) {
+    // 게시글 삭제 (본인 게시글만 삭제)
+    public void deletePost(Long postId, User user) {
         Post post = postRepository.findById(postId)
                 .orElseThrow(() -> new RuntimeException("Post not found with id " + postId));
+
+        if (!post.getUser().getUserId().equals(user.getUserId())) {
+            throw new RuntimeException("You are not authorized to delete this post");
+        }
+
         postRepository.delete(post);
     }
 }
